refactor(proxy): extract shared helpers into proxy_common (#2080)
The Caddy engine imports five pure helpers from the nginx engine
(proxy.py): detect_host_ipv4s, _is_loopback, _FALLBACK_ACL_SUBNETS,
_FALLBACK_DENY_SUBNETS, _proxy_preexec (plus their private deps
_LOOPBACK_NETS and the prctl machinery). Move them into a new
klangk.proxy_common so proxy.py can be deleted (#2081) without taking
the Caddy engine dependencies with it.

No behavior change: proxy.py re-exports the five, so all existing
`from klangk.proxy import ...` callers keep working; caddy.py now
imports them from proxy_common. Part 1/5 of the nginx retirement (#1642).
