+++ src/klangk/klangk/proxy.py
@@ -28,38 +28,27 @@ settings.
 from __future__ import annotations
 
 import asyncio
-import ctypes
-import ctypes.util
 import ipaddress
 import logging
 import os
 import signal
 import shutil
 import stat
-import subprocess
-import sys
 from pathlib import Path
 
-
-logger = logging.getLogger(__name__)
-
-# Loopback ranges excluded from the catch-all deny (CONTAINER_DENY) — local
-# browsers connect via loopback and must reach the full UI/API. Matches the
-# ``_is_loopback`` helper in the old nginx.sh (127.0.0.0/8 + ::1).
-_LOOPBACK_NETS = (
-    ipaddress.ip_network("127.0.0.0/8"),
-    ipaddress.ip_network("::1/128"),
+# Pure helpers shared with the Caddy engine live in :mod:`klangk.proxy_common`
+# (#1642). Re-exported here so existing ``from klangk.proxy import …`` callers
+# (and this module's own use) keep working until the nginx engine is removed.
+from klangk.proxy_common import (
+    _FALLBACK_ACL_SUBNETS,
+    _FALLBACK_DENY_SUBNETS,
+    _is_loopback,
+    _proxy_preexec,
+    detect_host_ipv4s,
 )
 
 
-def _is_loopback(addr: str) -> bool:
-    """True for any address in 127.0.0.0/8 or ::1."""
-    try:
-        ip = ipaddress.ip_address(addr)
-    except ValueError:
-        return False
-    return any(ip in net for net in _LOOPBACK_NETS)
-
+logger = logging.getLogger(__name__)
 
 # ---------------------------------------------------------------------------
 # Upstream constructors (pure — no settings)
@@ -84,32 +73,6 @@ def uds_upstream(socket_path: str) -> str:
 # ---------------------------------------------------------------------------
 # Environment probes (auto-detection, not settings)
 # ---------------------------------------------------------------------------
-
-
-def detect_host_ipv4s() -> list[str]:
-    """Auto-detect this host's IPv4 addresses (the pasta-NAT container source set).
-
-    Podman rootless default (pasta) shares the host network via userspace NAT,
-    so container traffic to ``host.containers.internal`` arrives from the
-    host's own IPv4. ``ip -4 addr show`` lists them (including 127.0.0.1 from
-    ``lo`` — wanted for CONTAINER_ACL, filtered out of CONTAINER_DENY below).
-    Returns ``[]`` on failure (caller falls back to RFC1918 ranges).
-    """
-    try:
-        out = subprocess.check_output(
-            ["ip", "-4", "addr", "show"], text=True, stderr=subprocess.DEVNULL
-        )
-    except (OSError, subprocess.CalledProcessError, FileNotFoundError):
-        return []
-    addrs: list[str] = []
-    for line in out.splitlines():
-        line = line.strip()
-        if line.startswith("inet "):
-            # "inet 192.168.1.5/24 ..."
-            token = line.split()[1]
-            ip = token.split("/")[0]
-            addrs.append(ip)
-    return addrs
 
 
 def stdout_is_reopenable() -> bool:
@@ -137,13 +100,6 @@ def stdout_is_reopenable() -> bool:
         # fd 1 unusable — assume reopenable to preserve legacy behavior.
         return True
     return not stat.S_ISSOCK(st.st_mode)
-
-
-# Fallback subnets when auto-detection yields nothing (mirrors nginx.sh):
-# 172.16/12 + 10/8 (common container ranges), explicitly NOT 192.168/16
-# (most common LAN range — allowing it would expose the LLM proxy to peers).
-_FALLBACK_ACL_SUBNETS = ["172.16.0.0/12", "10.0.0.0/8", "127.0.0.1"]
-_FALLBACK_DENY_SUBNETS = ["172.16.0.0/12", "10.0.0.0/8"]
 
 
 # ---------------------------------------------------------------------------
@@ -776,31 +732,6 @@ http {{
         if found:
             return found
         return "/usr/sbin/nginx"
-
-
-_PR_SET_PDEATHSIG = 1
-_HAS_PDEATHSIG = sys.platform == "linux"
-if _HAS_PDEATHSIG:  # pragma: no cover – linux-only
-    _libc = ctypes.CDLL(
-        ctypes.util.find_library("c") or "libc.so.6", use_errno=True
-    )
-
-
-def _proxy_preexec() -> None:  # pragma: no cover  – runs in forked child
-    """New session (for killpg) + auto-SIGTERM when parent dies (#1533).
-
-    ``os.setsid()`` puts nginx in its own process group so ``stop()`` can
-    ``os.killpg`` the entire tree on clean shutdown.
-
-    On Linux, ``prctl(PR_SET_PDEATHSIG, SIGTERM)`` asks the kernel to send
-    SIGTERM to the nginx master if klangkd dies without calling ``stop()``
-    (e.g. SIGKILL).  nginx handles SIGTERM by forwarding SIGQUIT to its
-    workers, so the whole tree exits.  macOS has no equivalent; on unclean
-    shutdown, orphaned nginx processes must be cleaned up externally.
-    """
-    os.setsid()
-    if _HAS_PDEATHSIG:
-        _libc.prctl(_PR_SET_PDEATHSIG, signal.SIGTERM)
 
 
 class ProxyWatchdog:
